pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-clickhouse_query_repair"
version = "0.1.0"
description = "Clickhouse Query Repair environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "clickhouse-connect>=0.7.0",
    "pip-system-certs>=5.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "openai>=1.0.0",
]

[project.scripts]
# Server entry point - enables running via: uv run --project . server
# or: python -m clickhouse_query_repair.server.app
server = "clickhouse_query_repair.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["clickhouse_query_repair", "clickhouse_query_repair.server"]
package-dir = { "clickhouse_query_repair" = ".", "clickhouse_query_repair.server" = "server" }

[tool.setuptools.package-data]
clickhouse_query_repair = ["tasks/*.json"]